TINY-8279: removed context from events and cleaned things up
