+++ modules/tinymce/src/core/test/ts/browser/keyboard/AutocompleterTest.ts
@@ -1,14 +1,12 @@
-import { Assertions, Keys, Waiter } from '@ephox/agar';
+import { Keys, Waiter } from '@ephox/agar';
 import { describe, it } from '@ephox/bedrock-client';
 import { Arr } from '@ephox/katamari';
-import { Hierarchy, SugarElement } from '@ephox/sugar';
-import { TinyContentActions, TinyHooks, TinySelections, TinyDom } from '@ephox/wrap-mcagar';
+import { TinyContentActions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 import { EditorEvent } from 'tinymce/core/api/PublicApi';
-import { AutocompleterEventArgs } from 'tinymce/core/autocomplete/AutocompleterEventArgs';
-import { AutocompleteContext, AutocompleteLookupData } from 'tinymce/core/autocomplete/AutocompleteTypes';
+import { AutocompleterEventArgs, AutocompleteLookupData } from 'tinymce/core/autocomplete/AutocompleteTypes';
 
 describe('browser.tinymce.core.keyboard.AutoCompleterTest', () => {
   const triggerChar = '+';
@@ -61,22 +59,6 @@ describe('browser.tinymce.core.keyboard.AutoCompleterTest', () => {
 
   const pWaitForEvents = (events: string[], expectedEvents: string[]) =>
     Waiter.pTryUntil('Waited for events to include expected events', () => assert.deepEqual(events, expectedEvents));
-
-  const assertRange = (editor: Editor, rng: Range, startPath: number[], startOffset: number, endPath: number[], endOffset: number) => {
-    const startContainer = Hierarchy.follow(TinyDom.body(editor), startPath).getOrDie();
-    const endContainer = Hierarchy.follow(TinyDom.body(editor), endPath).getOrDie();
-
-    Assertions.assertDomEq('Should be expected start container', startContainer, SugarElement.fromDom(rng.startContainer));
-    assert.equal(rng.startOffset, startOffset, 'Should be expected start offset');
-    Assertions.assertDomEq('Should be expected end container', endContainer, SugarElement.fromDom(rng.endContainer));
-    assert.equal(rng.endOffset, endOffset, 'Should be expected end offset');
-  };
-
-  const assertContext = (editor: Editor, context: AutocompleteContext, expectedText: string) => {
-    assert.equal(context.text, expectedText);
-    assert.equal(context.triggerChar, '+');
-    assertRange(editor, context.range, [ 0 ], 0, [ 0 ], 1);
-  };
 
   const asseertLookupData = (lookupData: AutocompleteLookupData, expectedMatchText: string) => {
     assert.equal(lookupData.columns, 1);
@@ -140,7 +122,7 @@ describe('browser.tinymce.core.keyboard.AutoCompleterTest', () => {
     editor.off('AutocompleterStart AutocompleterUpdate AutocompleterEnd', collect);
   });
 
-  it('autocompleter events start, update should have context', async () => {
+  it('autocompleter events start, update should have lookupData', async () => {
     const editor = hook.editor();
     const eventArgs: AutocompleterEventArgs[] = [];
     const collect = (args: EditorEvent<AutocompleterEventArgs>) => eventArgs.push(args);
@@ -148,21 +130,19 @@ describe('browser.tinymce.core.keyboard.AutoCompleterTest', () => {
     editor.on('AutocompleterStart AutocompleterUpdate', collect);
     await pOpenAutocompleter(editor);
 
-    await Waiter.pTryUntil('Waited for AutocompleterStart to include context/lookupData', () => {
+    await Waiter.pTryUntil('Waited for AutocompleterStart to include lookupData', () => {
       assert.equal(eventArgs.length, 1);
       assert.equal(eventArgs[0].lookupData.length, 1);
 
-      assertContext(editor, eventArgs[0].context, '');
       asseertLookupData(eventArgs[0].lookupData[0], '');
     });
 
     await pUpdateWithChar(editor, 'a');
 
-    await Waiter.pTryUntil('Waited for AutocompleterUpdate to include context/lookupData', () => {
+    await Waiter.pTryUntil('Waited for AutocompleterUpdate to include lookupData', () => {
       assert.equal(eventArgs.length, 2);
       assert.equal(eventArgs[1].lookupData.length, 1);
 
-      assertContext(editor, eventArgs[1].context, 'a');
       asseertLookupData(eventArgs[1].lookupData[0], 'a');
     });
